refactor(document): move document from internal to pkg
The package is a general-purpose typed YAML/JSON document reader-writer keyed on file extension. Its whole
exported surface is Read, ReadFile, Write and three options, and it imports nothing from this module -- stdlib
and yaml.v3 only. Nothing in it is devlore-, lore-, writ- or CLI-specific.

It was consumed from cmd/internal/lorepackage and most of the CLI, which first suggested moving it under
cmd/internal to sit beside its consumers. The code argues the other way. Three pkg/op/provider/plan tests use it
to round-trip a trace, and Go's own rule would make cmd/internal/document unimportable from pkg -- so those
consumers are not an accident to clean up, they are evidence the package is what it looks like. The placement
question was never internal-versus-cmd/internal; it was private convenience versus published utility, and this
is the latter.

It also becomes the codec for producing json, yaml and protobuf documents, which settles it: a codec is not CLI
support.

Nothing else changes. The move is a git mv plus an import-path rewrite across seventeen files, with no edit to
the package's contents -- deliberately, so the diff reads as a move. The doc comment still describes YAML and
JSON alone, and the third rendering will bring its own change.

Six consumers needed reformatting rather than none: pkg/document sorts to a different position than
internal/document within the NobleFactor import group, so the rewrite left their import blocks out of order.

Live references follow the move; docs/plans/structured-document-io.md does not. That plan records creating the
package under internal as complete, which was true when written, and rewriting it would falsify a history
rather than fix a pointer.
