+++ cmd/devlore-index/main.go
@@ -27,7 +27,7 @@ import (
 
 	"gopkg.in/yaml.v3"
 
-	"github.com/NobleFactor/devlore-cli/internal/document"
+	"github.com/NobleFactor/devlore-cli/pkg/document"
 )
 
 // KnowledgeIndex represents the index.yaml manifest for a knowledge domain.
