+++ cmd/internal/lorepackage/registry.go
@@ -18,8 +18,8 @@ import (
 
 	"github.com/NobleFactor/devlore-cli/cmd/internal/config"
 	"github.com/NobleFactor/devlore-cli/cmd/internal/devlore"
-	"github.com/NobleFactor/devlore-cli/internal/document"
 	"github.com/NobleFactor/devlore-cli/internal/registry"
+	"github.com/NobleFactor/devlore-cli/pkg/document"
 )
 
 // Registry provides access to a devlore registry.
