+++ cmd/writ/writ/migrate/execute.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/NobleFactor/devlore-cli/cmd/internal/cli"
-	"github.com/NobleFactor/devlore-cli/internal/document"
+	"github.com/NobleFactor/devlore-cli/pkg/document"
 	"github.com/NobleFactor/devlore-cli/pkg/fsroot"
 	"github.com/NobleFactor/devlore-cli/pkg/op"
 	"github.com/NobleFactor/devlore-cli/pkg/op/provider/file"
