+++ internal/credentials/file.go
@@ -7,7 +7,7 @@ import (
 	"errors"
 	"os"
 
-	"github.com/NobleFactor/devlore-cli/internal/document"
+	"github.com/NobleFactor/devlore-cli/pkg/document"
 	"github.com/NobleFactor/devlore-cli/pkg/xdg"
 )
